Refactor chess navigation into standalone component
Created a dedicated ChessNavigationComponent for chess-related menus and moved all related logic and template from ChessComponent. Simplified ChessComponent by delegating navigation responsibilities, improving separation of concerns and maintainability. Added unit tests for the new component.
