+++ website/src/app/chess/chess-event/chess-event-participants/chess-event-participants.component.html
@@ -1,4 +1,5 @@
 <p-card *ngIf="participants && participants.length != 0">
+  <ng-template #header></ng-template>
   <ng-template #title>{{'chess.event.participants' | translate}}</ng-template>
 
   <p-table stripedRows *ngIf="participants && participants.length != 0" [value]="participants">
