+++ website/src/app/chess/chess-event/chess-event.component.html
@@ -29,7 +29,11 @@
               </p-card>
             </div>
 
-            <app-chess-event-participants class="4k:col-span-3 xlScreen:col-span-2 lScreen:col-span-2 desktop:col-span-2 laptop:col-span-1 tablet:col-span-1 mobile:col-span-1" *ngIf="event" [event]="event"></app-chess-event-participants>
+            <app-chess-event-participants
+              class="4k:col-span-3 xlScreen:col-span-2 lScreen:col-span-2 desktop:col-span-2 laptop:col-span-1 tablet:col-span-1 mobile:col-span-1"
+              *ngIf="event"
+              [event]="event"
+            ></app-chess-event-participants>
           </div>
