+++ website/src/app/chess/chess-event/chess-event.component.scss
@@ -5,7 +5,3 @@ iframe{
   border: none; /* Remove default border */
   overflow: hidden;
 }
-
-:host ::ng-deep .p-card {
-  height: 100%;
-}
